requirements.txt
pyautogui==0.9.54
PyVirtualDisplay==3.0
python-xlib==0.33
